Take into account the new configuration library:  - modify the configuration to XML format  - enable using more than one file per binary  - use the new API

+++ platine-core/src/qos_data/bloc_ip_qos.h
@@ -131,7 +131,7 @@ class BlocIPQoS: public mgl_bloc
 	map < unsigned short, TrafficCategory * > categoryMap;
 
 	/// Category to be used when classifier returns no category
-	unsigned short defaultCategory;
+	long int defaultCategory;
 
 	// TUN file descriptor
 	int _tun_fd;
